[fix]Refactored code
