--- lucid-book-store/src/components/Auth/index.jsx
@@ -1,3 +1,0 @@
-export { Login } from "./login";
-export { Signup } from "./signup";
-export { Logout } from "./logout";
