--- lucid-book-store/src/components/Auth/login.css
@@ -1,19 +1,0 @@
-.login,
-.signup {
-  width: fit-content;
-}
-.view-password {
-  cursor: pointer;
-}
-.form-custom .form-input {
-  width: 100%;
-}
-.logout-wrapper {
-  width: 100vw;
-  height: 100vh;
-  z-index: 2;
-  position: fixed;
-}
-.logout .modal {
-  background-color: var(--background-white);
-}
