+++ lucid-book-store/src/components/index.jsx
@@ -1,10 +1,18 @@
-export { Homepage } from "./homepage/homepage";
+export {
+  AddToCartProductListing,
+  IncreaseProductQuantity,
+  DecreaseProductQuantity,
+  RemoveFromCart,
+} from "./buttons/cart-buttons";
 export { Navbar } from "./navbar/navbar";
 export { Footer } from "./footer/footer";
-export { ProductListing } from "./product-listing/product-listing";
-export { Cart } from "./cart-wishlist/cart";
-export { Wishlist } from "./cart-wishlist/wishlist";
-export { Login, Signup } from "./Auth";
-export { SearchResult } from "./navbar/search-results";
-export { SingleProductPage } from "./single-product-page/single-product-page";
-export { ErrorPage } from "./error-page/error-page";
+export {
+  AddToWishlistLarge,
+  AddToWishlistSmall,
+  RemoveFromWishlist,
+  AddToCartWishlist,
+  AddToWishlistSingleProductPage,
+} from "./buttons/wishlist-buttons";
+export { CartCard } from "./cards/cart-card";
+export { WishlistCard } from "./cards/wishlist-card";
+export { Logout } from "./modal/logout";
