+++ lucid-book-store/src/components/modal/modal.css
@@ -1,0 +1,9 @@
+.logout-wrapper {
+  width: 100vw;
+  height: 100vh;
+  z-index: 2;
+  position: fixed;
+}
+.logout .modal {
+  background-color: var(--background-white);
+}
