+++ lucid-book-store/src/components/navbar/navbar.css
@@ -10,7 +10,7 @@
   text-align: center;
 }
 .nav-search input {
-  width: 280rem;
+  width: 28rem;
 }
 .search-results {
   background-color: var(--background-white);
