+++ lucid-book-store/src/pages/cart/components/index.jsx
@@ -1,0 +1,2 @@
+export { CartProducts } from "./cart-products";
+export { CartSummary } from "./cart-summary";
